+++ packages/leaflet-map-adapter/src/LeafletMapAdapter.ts
@@ -10,9 +10,9 @@ import {
   LocateOptions,
   LocationEvents,
   Locate,
-  BaseMapEvents,
+  MainMapEvents,
 } from '@nextgis/webmap';
-import L, {
+import {
   Map,
   Control,
   Layer,
@@ -49,7 +49,7 @@ export class LeafletMapAdapter implements MapAdapter<Map, any, Control> {
     ATTRIBUTION: AttributionControl,
   };
 
-  static Map = L;
+  static Map = Map;
 
   options: MapOptions = { target: 'map' };
 
@@ -58,7 +58,7 @@ export class LeafletMapAdapter implements MapAdapter<Map, any, Control> {
   emitter = new EventEmitter();
   map?: Map;
 
-  private _universalEvents: (keyof BaseMapEvents)[] = [
+  private _universalEvents: (keyof MainMapEvents)[] = [
     'zoomstart',
     'zoom',
     'zoomend',