+++ packages/leaflet-map-adapter/src/index.ts
@@ -1,7 +1,3 @@
-import L from 'leaflet';
 import { LeafletMapAdapter } from './LeafletMapAdapter';
 
-export { LeafletMapAdapter };
 export default LeafletMapAdapter;
-
-export { L };