+++ packages/ngw-connector/src/utils/resourceToQuery.ts
@@ -1,9 +1,10 @@
-import { DeepPartial, isObject, defined } from '@nextgis/utils';
+import { DeepPartial, defined } from '@nextgis/utils';
 import { Resource } from '../types/ResourceItem';
+import { isObject } from './isObject';
 
 /**
  * { keyname, parent: { id }} > { keyname, parent__id }
- * @param resource any property from NGW resource item
+ * @param resource - Any property from NGW resource item
  */
 export function resourceToQuery(
   resource: DeepPartial<Resource>,