+++ packages/ngw-mapbox/src/index.ts
@@ -1,18 +1,13 @@
 import 'mapbox-gl/dist/mapbox-gl.css';
 
-import NgwMap, { NgwLayerOptions, NgwMapOptions } from '@nextgis/ngw-map';
+import { NgwMap, NgwLayerOptions, NgwMapOptions } from '@nextgis/ngw-map';
 import MapboxglMapAdapter, {
   MapboxglMapAdapterOptions,
 } from '@nextgis/mapboxgl-map-adapter';
 import { ResourceAdapter } from '@nextgis/ngw-kit';
 import { Map } from 'mapbox-gl';
 
-export class NgwMapbox extends NgwMap<
-  Map,
-  string[],
-  any,
-  MapboxglMapAdapterOptions
-> {
+class NgwMapbox extends NgwMap<Map, string[], any, MapboxglMapAdapterOptions> {
   constructor(options: NgwMapOptions) {
     super(new MapboxglMapAdapter(), options);
   }