+++ packages/vue-ngw-map/src/index.ts
@@ -1,7 +1,3 @@
-import { VueNgwMap } from './components/VueNgwMap';
-
+export * from './components/VueNgwMap';
 export * from './components/VueNgwResource';
 export * from './components/VueNgwControl';
-
-export { VueNgwMap };
-export default VueNgwMap;